feat: support animation mode in model viewer
